style-fix

+++ resources/views/admin/penduduk/rtm/index.blade.php
@@ -16,44 +16,63 @@
     <div class="box box-info">
         <div class="box-header with-border">
             @if (can('u'))
-                <a href="{{ ci_route('rtm.form') }}" title="Tambah" data-remote="false" data-toggle="modal"
-                    data-target="#modalBox" data-title="Tambah"
-                    class="btn btn-social bg-olive btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block"><i
-                        class='fa fa-plus'></i>Tambah</a>
+                <a
+                    href="{{ ci_route('rtm.form') }}"
+                    title="Tambah"
+                    data-remote="false"
+                    data-toggle="modal"
+                    data-target="#modalBox"
+                    data-title="Tambah"
+                    class="btn btn-social bg-olive btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block"
+                ><i class='fa fa-plus'></i>Tambah</a>
             @endif
             @if (can('h'))
-                <a href="#confirm-delete" title="Hapus Data"
-                    onclick="deleteAllBox('mainform','{{ ci_route('rtm.delete') }}')"
-                    class="btn btn-social btn-danger btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block hapus-terpilih"><i
-                        class='fa fa-trash-o'></i> Hapus</a>
+                <a href="#confirm-delete" title="Hapus Data" onclick="deleteAllBox('mainform','{{ ci_route('rtm.delete') }}')" class="btn btn-social btn-danger btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block hapus-terpilih"><i class='fa fa-trash-o'></i>
+                    Hapus</a>
             @endif
 
             @if (can('u'))
-            <a href="{{ ci_route('suplemen.impor') }}"
-                class="btn btn-social bg-navy btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block btn-import"
-                title="Impor Data" data-target="#impor" data-remote="false" data-toggle="modal" data-backdrop="false"
-                data-keyboard="false"><i class="fa fa-upload"></i>Impor</a>
+                <a
+                    href="{{ ci_route('suplemen.impor') }}"
+                    class="btn btn-social bg-navy btn-sm visible-xs-block visible-sm-inline-block visible-md-inline-block visible-lg-inline-block btn-import"
+                    title="Impor Data"
+                    data-target="#impor"
+                    data-remote="false"
+                    data-toggle="modal"
+                    data-backdrop="false"
+                    data-keyboard="false"
+                ><i class="fa fa-upload"></i>Impor</a>
             @endif
             <div class="btn-group-vertical">
-                <a class="btn btn-social bg-orange btn-sm" data-toggle="dropdown"><i
-                        class='fa fa-arrow-circle-down'></i> Laporan</a>
+                <a class="btn btn-social bg-orange btn-sm" data-toggle="dropdown"><i class='fa fa-arrow-circle-down'></i> Laporan</a>
                 <ul class="dropdown-menu" role="menu">
                     <li>
-                        <a href="{{ ci_route('rtm.ajax_cetak.cetak') }}"
-                            class="btn btn-social btn-block btn-sm" title="Cetak Data" data-remote="false"
-                            data-toggle="modal" data-target="#modalBox" data-title="Cetak Data"><i class="fa fa-print"></i>
+                        <a
+                            href="{{ ci_route('rtm.ajax_cetak.cetak') }}"
+                            class="btn btn-social btn-block btn-sm"
+                            title="Cetak Data"
+                            data-remote="false"
+                            data-toggle="modal"
+                            data-target="#modalBox"
+                            data-title="Cetak Data"
+                        ><i class="fa fa-print"></i>
                             Cetak</a>
                     </li>
                     <li>
-                        <a href="{{ ci_route('rtm.ajax_cetak.unduh') }}"
-                            class="btn btn-social btn-block btn-sm" title="Unduh Data" data-remote="false"
-                            data-toggle="modal" data-target="#modalBox" data-title="Unduh Data"><i
-                                class="fa fa-file-excel-o"></i> Unduh</a>
+                        <a
+                            href="{{ ci_route('rtm.ajax_cetak.unduh') }}"
+                            class="btn btn-social btn-block btn-sm"
+                            title="Unduh Data"
+                            data-remote="false"
+                            data-toggle="modal"
+                            data-target="#modalBox"
+                            data-title="Unduh Data"
+                        ><i class="fa fa-file-excel-o"></i> Unduh</a>
                     </li>
                 </ul>
-            </div>            
+            </div>
         </div>
-        <div class="box-body">            
+        <div class="box-body">
             <div class="row">
                 <div class="col-sm-2">
                     <select id="status" class="form-control input-sm select2">
@@ -74,7 +93,7 @@
                 <div class="col-sm-2">
                     <select id="dusun" class="form-control input-sm select2">
                         <option value="">Pilih Dusun</option>
-                        @foreach ($wilayah as $item)                            
+                        @foreach ($wilayah as $item)
                             <option value="{{ $item->id }}">{{ $item->dusun }}</option>
                         @endforeach
                     </select>
@@ -94,14 +113,14 @@
                 <div class="col-sm-2">
                     <select id="rt" class="form-control input-sm select2">
                         <option value="">Pilih RT</option>
-                        @foreach ($wilayah as $item)                            
+                        @foreach ($wilayah as $item)
                             @foreach ($item->rws as $child)
                                 <optgroup value={{ $child->id }} label="{{ $child->rw }}">
                                     @foreach ($item->rts as $rt)
                                         <option value="{{ $rt->id }}">{{ $rt->rt }}</option>
                                     @endforeach
                                 </optgroup>
-                            @endforeach    
+                            @endforeach
                         @endforeach
                     </select>
                 </div>
@@ -109,9 +128,9 @@
             <hr>
             {!! form_open(null, 'id="mainform" name="mainform"') !!}
             @if ($judul_statistik)
-				<h5 class="box-title text-center"><b>{{ $judul_statistik }}</b></h5>
-			@endif
-            <div class="table-responsive">                
+                <h5 class="box-title text-center"><b>{{ $judul_statistik }}</b></h5>
+            @endif
+            <div class="table-responsive">
                 <table class="table table-bordered table-hover" id="tabeldata">
                     <thead>
                         <tr>
@@ -137,7 +156,7 @@
         </div>
     </div>
 
-    @include('admin.layouts.components.konfirmasi_hapus')    
+    @include('admin.layouts.components.konfirmasi_hapus')
     @include('admin.penduduk.rtm.impor')
     @include('admin.penduduk.rtm.dtks_modal')
 @endsection
@@ -170,11 +189,11 @@
                         req.dusun = $('#dusun').val();
                         req.rw = $('#rw').val();
                         req.rt = $('#rt').val();
-                        if(filterColumn['status']){
+                        if (filterColumn['status']) {
                             req.bdt = filterColumn['status'];
                         }
                     }
-                },                
+                },
                 columns: [{
                         data: 'ceklist',
                         class: 'padat',
@@ -205,10 +224,10 @@
                         name: 'no_kk',
                         searchable: true,
                         orderable: true,
-                        render: function(row, data, item){
+                        render: function(row, data, item) {
                             return `<span class="text-bold">${item.no_kk}</span>`;
                         }
-                    },                    
+                    },
                     {
                         data: 'kepala_keluarga.nama',
                         name: 'kepalaKeluarga.nama',
@@ -232,7 +251,7 @@
                         name: 'anggota_count',
                         searchable: false,
                         orderable: false
-                    },                    
+                    },
                     {
                         data: 'kepala_keluarga.alamat_wilayah',
                         name: 'alamat_wilayah',
@@ -268,7 +287,7 @@
                         orderable: true
                     },
 
-                    
+
                 ],
                 order: [
                     [4, 'asc']
@@ -283,7 +302,7 @@
                 TableData.column(2).visible(false);
             }
 
-            $('#dusun').change(function(){
+            $('#dusun').change(function() {
                 let _label = $(this).find('option:selected').text()
                 $('#rw').find(`optgroup`).prop('disabled', 1)
                 if ($(this).val()) {
@@ -293,28 +312,28 @@
                     $('#rw').closest('div').hide()
                     $('#rw').find(`optgroup`).prop('disabled', 1)
                 }
-                $('#rw').val('')                
+                $('#rw').val('')
                 $('#rw').trigger('change')
             })
 
-            $('#rw').change(function(){
+            $('#rw').change(function() {
                 let _label = $(this).find('option:selected').val()
-                $('#rt').find(`optgroup`).prop('disabled', 1)                
+                $('#rt').find(`optgroup`).prop('disabled', 1)
                 if ($(this).val()) {
                     $('#rt').closest('div').show()
-                    $('#rt').find(`optgroup[value="${_label}"]`).prop('disabled', 0)                    
+                    $('#rt').find(`optgroup[value="${_label}"]`).prop('disabled', 0)
                 } else {
                     $('#rt').closest('div').hide()
                     $('#rt').find(`optgroup`).prop('disabled', 1)
-                }               
+                }
             })
 
             $('#status, #jenis_kelamin, #dusun, #rw, #rt').change(function() {
                 TableData.draw()
             })
 
-            if(filterColumn){
-                if (filterColumn['sex'] > 0){
+            if (filterColumn) {
+                if (filterColumn['sex'] > 0) {
                     $('#jenis_kelamin').val(filterColumn['sex'])
                 }
             }
